# openapi/hub.yaml
openapi: 3.1.0
info:
  title: Agentic GPT Hub API
  version: 0.3.1
servers:
  - url: https://agentic-gpt.example.com
security:
  - bearerAuth: []
paths:
  /v1/info:
    get:
      operationId: getInfo
      description: Return Hub runtime metadata, public base URL, remote confirmation status, and safe aggregate counts.
      x-openai-isConsequential: false
      responses:
        "200":
          description: Safe Hub runtime information.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubInfoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/agents:
    get:
      operationId: listAgents
      description: List registered local agents, online status, capabilities, and safe config summaries. Use this before choosing an agentId.
      x-openai-isConsequential: false
      responses:
        "200":
          description: Enabled agents with safe configuration summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAgentsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/events:
    get:
      operationId: listEvents
      description: List one explicitly selected enabled Agent's persisted inbox. Defaults to pending events and 20 items; hidden pending events remain listable. The response includes the Agent's compact events panel when available. Offline, timeout, and Hub errors omit events.
      x-openai-isConsequential: false
      parameters:
        - name: agentId
          in: query
          required: true
          description: Target enabled local Agentic agent id; this API never selects or merges agents.
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EventStatus"
            default: pending
        - name: severity
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EventSeverity"
        - name: limit
          in: query
          required: false
          description: Defaults to 20; the Agent clamps the effective page size to 1 through 100.
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned by a prior listEvents response, scoped to the same Agent and filters.
          schema:
            type: string
      responses:
        "200":
          description: Flat event list DTO plus the optional compact Agent events panel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
              examples:
                page:
                  summary: Pending events and compact panel
                  value:
                    items:
                      - eventId: evt_example
                        summary: Process completed
                        severity: low
                        createdAt: "2026-10-01T12:00:00Z"
                        status: pending
                    events:
                      current: "low: 1 | medium: 0 | high: 0"
                      new:
                        - "evt_example | Process completed": "low | 2026-10-01T12:00:00Z"
        "400":
          $ref: "#/components/responses/AgentError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AgentError"
        "500":
          $ref: "#/components/responses/AgentError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/events/{eventId}:
    get:
      operationId: getEvent
      description: Read one full event record for the explicitly selected enabled Agent. Reading does not mark it handled. The response includes the Agent's compact events panel when available; offline, timeout, and Hub errors omit events.
      x-openai-isConsequential: false
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
        - name: agentId
          in: query
          required: true
          description: Target enabled local Agentic agent id; event IDs are never looked up across agents.
          schema:
            type: string
      responses:
        "200":
          description: Full flat event record plus the optional compact Agent events panel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventGetResponse"
              examples:
                record:
                  summary: Event record with its source and compact panel
                  value:
                    eventId: evt_example
                    message: Process completed successfully
                    severity: low
                    createdAt: "2026-10-01T12:00:00Z"
                    status: pending
                    source:
                      kind: process
                      ref: process-1
                    shownCount: 1
                    expiresAt: "2026-10-02T12:00:00Z"
                    events:
                      current: "low: 1 | medium: 0 | high: 0"
                      new:
                        - "evt_example | Process completed": "low | 2026-10-01T12:00:00Z"
        "400":
          $ref: "#/components/responses/AgentError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AgentError"
        "500":
          $ref: "#/components/responses/AgentError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/events/mark:
    post:
      operationId: markEvents
      description: Idempotently mark selected events in one explicitly selected enabled Agent's inbox as handled. Unknown IDs are returned in notFoundIds; this changes only inbox handling state. The response includes the Agent's compact events panel when available; offline, timeout, and Hub errors omit events.
      x-openai-isConsequential: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventMarkRequest"
            examples:
              mark:
                summary: Mark a batch in one Agent inbox
                value:
                  agentId: agent-example
                  eventIds: [evt_example, evt_other]
      responses:
        "200":
          description: Handled and unknown event IDs, plus the optional compact Agent events panel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventMarkResponse"
              examples:
                marked:
                  summary: Per-ID result
                  value:
                    handledIds: [evt_example]
                    notFoundIds: [evt_other]
                    events:
                      current: "low: 0 | medium: 0 | high: 0"
                      new: []
        "400":
          $ref: "#/components/responses/AgentError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AgentError"
        "500":
          $ref: "#/components/responses/AgentError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/process/exec:
    post:
      operationId: processExec
      description: Start one managed process and wait briefly. Complete inline output and result are included only when the serialized creation response fits within 8192 bytes; overflow output uses a shared stdout/stderr preview capped at 2048 bytes. Use process status/output/result/cancel routes for lifecycle follow-up.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessExecRequest"
      responses:
        "200":
          description: Managed process creation status with bounded inline completion evidence or a typed validation/admission error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessExecResponse"
                  - $ref: "#/components/schemas/AgentErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/process/batch:
    post:
      operationId: processBatch
      description: Atomically admit multiple managed processes under one confirmation decision. The full batch response shares an 8192-byte inline budget and bounded output previews; use process status/output/result/cancel routes for lifecycle control.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessBatchExecRequest"
      responses:
        "200":
          description: Managed process batch status and ordered child process responses, or a typed validation/admission error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessBatchResponse"
                  - $ref: "#/components/schemas/AgentErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/process:
    get:
      operationId: listProcesses
      description: List active or recently retained managed processes for one local Agentic agent with optional group/kind/state filters and cursor pagination. If the Agent is unavailable, cached status metadata is returned with explicit freshness; cursors issued by the Agent cannot be continued from cache.
      x-openai-isConsequential: false
      parameters:
        - name: agentId
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProcessKind"
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProcessState"
        - name: group
          in: query
          required: false
          description: Exact human-readable workstream filter.
          schema:
            type: string
            minLength: 1
            maxLength: 32
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned by a prior process.list response.
          schema:
            type: string
      responses:
        "200":
          description: Active or recently retained process status metadata, with freshness.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessListResponse"
                  - $ref: "#/components/schemas/ProcessLiveErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Agent unavailable while continuing an Agent-issued cursor; no cache-generated cursor is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessListUnavailableResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/process/{processId}:
    get:
      operationId: getProcessStatus
      description: Inspect or briefly wait for managed process status only. Status never includes stdout, stderr, or result data; cached metadata is explicitly marked when the Agent is unavailable.
      x-openai-isConsequential: false
      parameters:
        - name: processId
          in: path
          required: true
          schema:
            type: string
        - name: agentId
          in: query
          required: true
          schema:
            type: string
        - name: waitSeconds
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 30
            default: 5
      responses:
        "200":
          description: Current process status, or explicitly cache-only status when the Agent is unavailable.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessStatusResponse"
                  - $ref: "#/components/schemas/ProcessCacheFallbackResponse"
                  - $ref: "#/components/schemas/ProcessLiveErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: Agent unavailable without a cached process status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessUnavailableResponse"
  /v1/process/{processId}/output:
    get:
      operationId: getProcessOutput
      description: Read a non-consuming output page from both process streams. Cursor offsets are raw-byte coordinates; the response preserves invalid UTF-8 with base64 encoding and explicitly reports retention gaps and EOF.
      x-openai-isConsequential: false
      parameters:
        - name: processId
          in: path
          required: true
          schema:
            type: string
        - name: agentId
          in: query
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          description: Opaque process-bound cursor for two absolute raw-byte stream offsets; omission starts from zero and reports any retention gaps.
          schema:
            type: string
        - name: maxBytes
          in: query
          required: false
          description: Maximum aggregate encoded bytes in the page; defaults to 8192 and is capped at 32768.
          schema:
            type: integer
            minimum: 1
            maximum: 32768
            default: 8192
      responses:
        "200":
          description: Lossless stdout and stderr page with cursor, gap, capture, and EOF evidence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessOutputResponse"
        "400":
          $ref: "#/components/responses/AgentError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AgentError"
        "500":
          $ref: "#/components/responses/AgentError"
        "503":
          description: Agent unavailable; output is never synthesized from Hub metadata cache.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessUnavailableResponse"
  /v1/process/{processId}/result:
    get:
      operationId: getProcessResult
      description: Read an explicitly retained structured result or an accurate unavailable status. Hub metadata cache never supplies process result content, and oversized or unretained results are never represented as partial JSON.
      x-openai-isConsequential: false
      parameters:
        - name: processId
          in: path
          required: true
          schema:
            type: string
        - name: agentId
          in: query
          required: true
          schema:
            type: string
        - name: maxBytes
          in: query
          description: Maximum retained result payload bytes to include; defaults to 8192 and is capped at 524288.
          schema:
            type: integer
            minimum: 1
            maximum: 524288
            default: 8192
      responses:
        "200":
          description: Retained structured result or explicit unavailable result status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResultResponse"
        "400":
          $ref: "#/components/responses/AgentError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AgentError"
        "500":
          $ref: "#/components/responses/AgentError"
        "503":
          description: Agent unavailable; result content is never synthesized from Hub metadata cache.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessUnavailableResponse"
  /v1/process/{processId}/cancel:
    post:
      operationId: cancelProcess
      description: Request kind-aware cancellation for one managed process and return observed termination evidence.
      x-openai-isConsequential: true
      parameters:
        - name: processId
          in: path
          required: true
          schema:
            type: string
        - name: agentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process cancellation outcome and observed termination evidence.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessCancelResponse"
                  - $ref: "#/components/schemas/ProcessLiveErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          description: Hub could not deliver the cancellation request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessUnavailableResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tmux/sessions:
    get:
      operationId: tmuxListSessions
      description: List persistent tmux shared workspaces. Start tmux workflows here, then use tmuxListPanes to locate the relevant pane.
      x-openai-isConsequential: false
      parameters:
        - name: agentId
          in: query
          required: true
          description: Target enabled local Agentic agent id.
          schema:
            type: string
      responses:
        "200":
          description: tmux sessions or a structured agent-side error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TmuxListSessionsResponse"
                  - $ref: "#/components/schemas/AgentErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/tmux/panes:
    get:
      operationId: tmuxListPanes
      description: List tmux panes and their cwd, foreground command, process/mode state, and isLikelyShell. Use tmuxExec for shell panes and tmuxPasteText for non-shell panes.
      x-openai-isConsequential: false
      parameters:
        - name: agentId
          in: query
          required: true
          description: Target enabled local Agentic agent id.
          schema:
            type: string
        - name: session
          in: query
          required: false
          description: Optional tmux session name to scope pane listing.
          schema:
            type: string
      responses:
        "200":
          description: Matching tmux panes or a structured agent-side error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TmuxListPanesResponse"
                  - $ref: "#/components/schemas/AgentErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/tmux/capture:
    post:
      operationId: tmuxCapturePane
      description: Capture recent pane history before acting or to observe progress after tmuxExec. This is read-only.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TmuxCaptureRequest"
      responses:
        "200":
          description: Pane capture or a structured agent-side error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmuxCaptureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/tmux/exec:
    post:
      operationId: tmuxExec
      description: Submit one structured program and argument vector atomically to a tmux shell pane. Pane cwd, command policy, path policy, confirmation, and audit apply. Returns a bounded post-submit pane snapshot after waitMs. Submission does not prove completion; use exec when an exit status is required.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TmuxExecRequest"
      responses:
        "200":
          description: Submission result or a structured policy/agent-side error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmuxOperationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/tmux/paste:
    post:
      operationId: tmuxPasteText
      description: Paste text into a non-shell tmux pane or TUI. Shell panes are rejected and must use tmuxExec. Confirmation defaults to true.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TmuxPasteRequest"
      responses:
        "200":
          description: Paste result or a structured agent-side error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmuxOperationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/tmux/sessions/create:
    post:
      operationId: tmuxCreateSession
      description: Create an idempotent persistent tmux workspace when no suitable session exists. Prefer reusing the default agentic session. cwd is checked by local path policy.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TmuxCreateSessionRequest"
      responses:
        "200":
          description: Session creation result or a structured agent-side error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmuxOperationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
  /v1/notify/channels:
    get:
      operationId: userNotifyChannels
      description: List Hub-native user notification channels. This does not use the active Room Agent.
      x-openai-isConsequential: false
      responses:
        "200":
          description: Available and unavailable notification channels.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotifyChannelsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Invalid or missing GPT Actions API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    GatewayTimeout:
      description: Hub timed out waiting for the local agent.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AgentError:
      description: Agent returned an error for a targeted online operation. The compact events panel is optional and is absent when the Hub or Agent could not retrieve it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentErrorResponse"
    InternalError:
      description: Hub internal error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ListAgentsResponse:
      type: object
      required:
        - agents
      additionalProperties: false
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentSummary"
...
    EventSeverity:
      type: string
      enum: [low, medium, high]
    EventStatus:
      type: string
      enum: [pending, handled, expired]
...
    EventListResponse:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventListItem"
        nextCursor:
          type: string
        events:
          $ref: "#/components/schemas/EventPanel"
...